move App into index.js
