+++ src/index.js
@@ -1,6 +1,23 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from 'App.js'
+import { ApolloProvider } from 'react-apollo'
+import { BrowserRouter as Router } from 'react-router-dom'
+import { useMediaPredicate } from 'react-media-hook'
+import apolloClient from 'apolloClient'
+import PrimaryLayout from 'pages/PrimaryLayout'
+import ScreenWidthContext from 'contexts/screenWidth'
 import './index.css'
+
+export function App () {
+  const isWide = useMediaPredicate('(min-width: 550px)')
+
+  return <ApolloProvider client={apolloClient}>
+    <Router>
+      <ScreenWidthContext.Provider value={isWide}>
+        <PrimaryLayout />
+      </ScreenWidthContext.Provider>
+    </Router>
+  </ApolloProvider>
+}
 
 ReactDOM.render(<App />, document.getElementById('root'))
